requirements.txt
boto3==1.26.148
pyautogui==0.9.54
python-dotenv==0.19.0
pycurl==7.45.2
opencv-python==4.7.0.72
paramiko==3.2.0
